App-Backend/requirements.txt
sqlalchemy
alembic
psycopg2-binary
python-dotenv
pydantic
fastapi
uvicorn